change module system to use floats instead of doubles

+++ src/main/java/com/github/lehjr/mpalib/client/gui/frame/IGuiFrame.java
@@ -76,7 +76,7 @@ public interface IGuiFrame extends IRect {
      * @param mouseX
      * @param mouseY
      */
-    void update(float mouseX, float mouseY);
+    void update(double mouseX, double mouseY);
 
     /**
      * Render elements of this frame. Ordering is important.
